Related to pull request #294.
When LDAP is being used with a bind_dn and a password provided in the configuration; a user can simply enter a valid username and leave the password field blank and still log in. This patch corrects that behavior and throws and AuthenticationError when the user does not provide a password.