Report missing credentials to the Rails application.
If no username or password is provided a MissingCredentialsError is
raised which causes a Rack caught exception and a 500 Error in gitlab.
Omniauth provides a way to raise such errors to the application by
using the 'fail!' method to pass the exception to the registered
failure handler. For gitlab this is the omniauth_controller code.

This is required to resolve gitlab issue #1077.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>