Report the reason for bind failure
When an account's credentials are correct, but they still cannot
bind it is usually due to a locked account.

This shouldn't give any extra info to an attacker (they must
already know the username and password) but it does help
users who get locked out (too many incorrect passwords, for example)
to know what's wrong.